Assignment03/Image_resizer/requirements.txt
flask==2.0.2
Pillow==8.3.1